Move icon helper back to Litter-Robot component and isoformat times on time sensors
